+++ src/IAmDifferent/IAmDifferentForm.css
@@ -1,4 +1,4 @@
-/* .mainContents {
+.mainContents {
   margin-top: 50px;
   margin-bottom:50px;
   font-family: 'adonis-web';
@@ -6,8 +6,8 @@
 
 
 input[type="checkbox"]{
-  width: 30px; 
-  height: 30px; 
+  width: 30px;
+  height: 30px;
   cursor: pointer;
 }
 
@@ -20,15 +20,14 @@ input[type="checkbox"]{
   height: auto;
 }
 
-.form-section-title{
+/* .form-section-title{
   background-color: #D9C739 !important;
   margin-bottom:2% !important;
   padding-top: 0.5% !important;
   padding-bottom:0.5% !important;
   font-weight: bold !important;
   font-size: 45px !important;
-
-}
+} */
 .form_field_text{
   font-size:30px;
 }
@@ -45,9 +44,9 @@ input[type="checkbox"]{
 #hint{
   color: #707070;
 }
-.options{
+/* .options{
   font-size: 40px;
-}
+} */
 
 @media screen and (max-width: 767px) {
 
@@ -58,20 +57,20 @@ input[type="checkbox"]{
     font-size: 45px;
   }
 
-  .formLabel {
+  /* .formLabel {
     font-size:25px;
-  }
+  } */
 
   .formConsent {
     font-size:25px;
   }
-  .form-section-title{
+  /* .form-section-title{
     background-color: "#D9C739" !important;
     margin-bottom:"2%" !important;
     padding-top:"0.5%" !important;
     padding-bottom:"0.5%" !important;
     font-weight: "bold" !important;
-  }
+  } */
   .hed{
      background-color: #D9C739;
      margin-bottom:2%;
@@ -91,9 +90,9 @@ input[type="checkbox"]{
   .contentsInParagraph {
     font-size: 45px;
   }
-  .formLabel {
+  /* .formLabel {
     font-size:20px;
-  }
+  } */
 
   .formConsent {
     font-size:25px;
@@ -113,14 +112,14 @@ input[type="checkbox"]{
     font-size:40px;
     margin: auto;
   }
-  .form-section-title{
+  /* .form-section-title{
     background-color: #D9C739 !important;
     margin-bottom:2% !important;
     padding-top: 0.5% !important;
     padding-bottom:0.5% !important;
     font-weight: bold !important;
     font-size: 45px !important;
-  }
+  } */
   .hed{
      background-color: #D9C739;
      margin-bottom:2%;
@@ -146,9 +145,9 @@ input[type="checkbox"]{
     font-size: 20px;
   }
 
-  .formLabel {
+  /* .formLabel {
     font-size:18px;
-  }
+  } */
 
   .formConsent {
     font-size:18px;
@@ -157,18 +156,18 @@ input[type="checkbox"]{
     font-size:40px ;
   }
   input[type="checkbox"]{
-    width: 25px; 
-    height: 25px; 
+    width: 25px;
+    height: 25px;
     cursor: pointer;
   }
-  .form-section-title{
+  /* .form-section-title{
     background-color: #D9C739 !important;
     margin-bottom:2% !important;
     padding-top: 0.5% !important;
     padding-bottom:0.5% !important;
     font-weight: bold !important;
     font-size: 45px !important;
-  }
+  } */
   .hed{
      background-color: #D9C739;
      margin-bottom:2%;
@@ -190,9 +189,9 @@ input[type="checkbox"]{
     font-size: 25px;
   }
 
-  .formLabel {
+  /* .formLabel {
     font-size:20px;
-  }
+  } */
 
   .formConsent {
     font-size:20px !important;
@@ -205,14 +204,14 @@ input[type="checkbox"]{
   .form-file-input{
     font-size: 25px;
   }
-  .form-section-title{
+  /* .form-section-title{
     background-color: #D9C739 !important;
     margin-bottom:2% !important;
     padding-top: 0.5% !important;
     padding-bottom:0.5% !important;
     font-weight: bold !important;
     font-size: 30px !important;
-  }
+  } */
   .submit_button{
        padding-left:  5%!important;
        padding-right: 5%!important;
@@ -223,15 +222,15 @@ input[type="checkbox"]{
       margin: auto !important;
   }
   input[type="checkbox"]{
-    width: 25px; 
-    height: 25px; 
+    width: 25px;
+    height: 25px;
     cursor: pointer;
   }
-  .form-section-title{
+  /* .form-section-title{
     font-family: "sofia-pro", Sans-serif;
     margin-bottom: -0.5%;
     font-size: 20px;
-  }
+  } */
   .hed{
      background-color: #D9C739;
      margin-bottom:2%;
@@ -240,4 +239,4 @@ input[type="checkbox"]{
     font-size: 20px!important;
     font-weight:bold;
   }
-} */
+}
