+++ src/IAmDifferent/i_am_different.js
@@ -1,12 +1,12 @@
 import React from "react";
-import newsletterContactUs from "./newsletterContactUs";
+// import newsletterContactUs from "./newsletterContactUs";
 import CardGame from "./cardGame";
 import styled from "@emotion/styled";
 import different_img from "../assets/Different.jpg";
 import { Container, Row, Col, Button } from "react-bootstrap";
 import ProgramInfo from "../programInfo";
 import "./i_am_different.css";
-import Header from "../header";
+// import Header from "../header";
 import { Link } from "react-router-dom";
 import NewsletterContactUs from "./newsletterContactUs";
 import Footer from "../footer-temp";
